add proof of lemma sqrt_le_J
added necessary hypotheses to lemma sqrt_le_J and added proof to address issue 228.

As on page 118 of [listdecoding] the argument x or "delta" in the function is the relative distance which should be between 0 and 1. On the other hand q is an integer as specified on page 117 it must be greater than 1 for the expressions to be defined. We must assume that q/(q-1) * x is less than 1 or else the expressions found in the J function will be undefined.